+++ core-ng/src/test/java/core/framework/internal/http/ServiceUnavailableInterceptorTest.java
@@ -6,13 +6,10 @@ import okhttp3.Protocol;
 import okhttp3.Request;
 import okhttp3.Response;
 import okhttp3.internal.connection.RealConnection;
-import okhttp3.internal.connection.RealConnectionPool;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.lang.reflect.Field;
-import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -31,8 +28,8 @@ class ServiceUnavailableInterceptorTest {
     }
 
     @Test
-    void intercept() throws IOException, NoSuchFieldException, IllegalAccessException {
-        var connection = new RealConnection(new RealConnectionPool(0, 10, TimeUnit.SECONDS), null);
+    void intercept() throws IOException {
+        var connection = new RealConnection(null, null);
         var request = new Request.Builder().url("http://localhost").build();
         var response = new Response.Builder().request(request)
                                              .protocol(Protocol.HTTP_2)
@@ -43,10 +40,6 @@ class ServiceUnavailableInterceptorTest {
         when(chain.connection()).thenReturn(connection);
 
         interceptor.intercept(chain);
-        Field field = RealConnection.class.getDeclaredField("noNewExchanges");
-
-        if (!field.trySetAccessible())
-            throw new Error("failed to access noNewExchanges field");
-        assertThat((boolean) field.get(connection)).isTrue();
+        assertThat(connection.noNewStreams).isTrue();
     }
 }