+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## Change log
-### 6.12.2 (4/2/2019 - )
+### 6.12.2 (4/2/2019 - 4/9/2019)
 * http: update undertow to 2.0.20
 * search: update es to 6.7.1
 * test: added assertEnumClass(class1).hasAllConstantsOf(class2)
 * http: update http().bean() to accept only one class (for simplification), support register enum
 * http: revert okHTTP to 3.12.2 due to https://github.com/square/okhttp/issues/4875
+* log: tweak trace log, still attempt to show warning if reached max trace length (to limit overall trace within 900k)
 
 ### 6.12.1 (4/1/2019)
 * kafka: fix kafka producer metrics can be NaN